requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=10.0.0
opencv-python>=4.8.0
scipy>=1.11.0

# SAM models
segment-anything>=1.0
transformers>=4.35.0

# 3D reconstruction
open3d>=0.18.0
trimesh>=4.0.0
pytorch3d>=0.7.5

# Video processing
ffmpeg-python>=0.2.0
moviepy>=1.0.3

# Utilities
tqdm>=4.66.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0

# API and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Logging and monitoring
loguru>=0.7.0
prometheus-client>=0.18.0

# Image processing
scikit-image>=0.21.0
albumentations>=1.3.0

